refactored all chart settings
