+++ viewpoint-frontend/src/components/chart/ChartArea.js
@@ -1,17 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
-import { Container } from '@mui/material';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Grid from '@mui/material/Grid';
-import TextField from '@mui/material/TextField';
-import Select from '@mui/material/Select';
-import MenuItem from '@mui/material/MenuItem';
-import InputLabel from '@mui/material/InputLabel';
-import OutlinedInput from '@mui/material/OutlinedInput';
-import FormControl from '@mui/material/FormControl';
-import { httpRequest } from '../../service/httpRequest';
-import { width } from '@mui/system';
 import TableArea from './Table/TableArea';
 import PieArea from './pie/PieArea';
 import LineArea from './line/LineArea';
@@ -19,11 +8,10 @@ import BarArea from './bar/BarArea';
 import Paper from '@mui/material/Paper';
 
 const ChartArea = (props) => {
-    const { chartData, chartResult } = props;
-    if (!chartData) return;
+    const { chartData } = props;
 
     return(
-        // <Box>
+        chartData && 
         <Box 
             height={"100%"} 
             bgcolor={"white"} 
@@ -36,14 +24,12 @@ const ChartArea = (props) => {
             <Typography variant="h4" fontWeight={600}>{chartData.name}</Typography>
                 <Chart {...props}/>
         </Box>
-        //{ </Box> }
     )
 }
 
 const Chart = (props) => {
     const { chartData } = props;
 
-    if (!chartData.data) return;
     if (chartData.chartType.toLowerCase() === "table") {
         return (<TableArea {...props}/>);
     } else if (chartData.chartType.toLowerCase() === "pie") {
