+++ viewpoint-frontend/src/components/chart/bar/BarArea.js
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { BarChart } from '@mui/x-charts/BarChart';
-import { Box, height, width } from '@mui/system';
 import dayjs from 'dayjs';
 
 const legendPlacement = {
@@ -19,25 +18,25 @@ const legendPlacement = {
     },
   };
 
-const BarArea = ({chartData, chartResult}) => {
-    if (!chartResult || !chartData) return;
-
-    const { rows, columns, data } = chartResult;
-    const { chartSettings, dataset } = chartData;
-    const { metrics, dimensions, xAxis } = chartSettings;
+const BarArea = (props) => {
+    const {chartData} = props;
+    const { chartSettings, rows, columns, data } = chartData;
+    const { metrics, dimensions, xAxis, xAxisType } = chartSettings;
 
     // Устаноить тип оси
     const sortedKeys = columns.sort();
-    const xColumn = dataset.columns.find(c => c.name == xAxis);
+    const xColumn = columns.find(c => c.name == xAxis);
+
     const xAxisSettings = {
         data: sortedKeys,
     };
-    if (xColumn["type"] === "Timestamp") {
+
+    if (xAxisType === "Timestamp") {
         xAxisSettings["data"] = sortedKeys.map(stringDate => new Date(stringDate));
         xAxisSettings["scaleType"] = 'time';
         xAxisSettings["tickMinStep"] = 3600 * 1000 * 24; // 24 h
         xAxisSettings["valueFormatter"] = (value) => (dayjs(value).format("YYYY-MM-DD"));
-    } else if (xColumn["type"] === "String") {
+    } else if (xAxisType === "String") {
         xAxisSettings["scaleType"] = 'band';
     }
